♻️ : move math.cuvers to math.interpolate
