+++ tactics2d/math/__init__.py
@@ -1,4 +1,4 @@
 from .geometry import Circle, Vector
-from .curves.bezier import Bezier
+from .interpolate.bezier import Bezier
 
 __all__ = ["Bezier", "Circle", "Vector"]
