--- tactics2d/math/curves/b_spline.py
@@ -1,6 +1,0 @@
-from .curve_base import CurveBase
-
-
-class BSpline(CurveBase):
-    def __init__(self) -> None:
-        super().__init__()
