+++ tactics2d/math/interpolate/__init__.py
