+++ tactics2d/math/interpolate/b_spline.py
@@ -1,0 +1,18 @@
+import numpy as np
+
+
+class BSpline:
+    """This class implements a B-spline curve interpolator.
+
+    Attributes:
+
+    """
+
+    def __init__(self, order, type):
+        self.order = order
+
+    def de_boor(self):
+        return
+
+    def get_curve(self, control_points: np.ndarray):
+        return
